reorganize things a bit and fix incorrect API usage
