+++ src/pages/ReimbursementAccount/BankAccountStep.js
@@ -1,0 +1,178 @@
+import _ from 'underscore';
+import React from 'react';
+import {View, Image} from 'react-native';
+import HeaderWithCloseButton from '../../components/HeaderWithCloseButton';
+import MenuItem from '../../components/MenuItem';
+import {Paycheck, Bank, Lock} from '../../components/Icon/Expensicons';
+import styles from '../../styles/styles';
+import TextLink from '../../components/TextLink';
+import Button from '../../components/Button';
+import Icon from '../../components/Icon';
+import colors from '../../styles/colors';
+import Navigation from '../../libs/Navigation/Navigation';
+import CONST from '../../CONST';
+import TextInputWithLabel from '../../components/TextInputWithLabel';
+import AddPlaidBankAccount from '../../components/AddPlaidBankAccount';
+import CheckboxWithLabel from '../../components/CheckboxWithLabel';
+import withLocalize, {withLocalizePropTypes} from '../../components/withLocalize';
+import exampleCheckImage from '../../../assets/images/example-check-image.png';
+import Text from '../../components/Text';
+
+const propTypes = {
+    ...withLocalizePropTypes,
+};
+
+class BusinessBankAccountNewPage extends React.Component {
+    constructor(props) {
+        super(props);
+
+        this.toggleTerms = this.toggleTerms.bind(this);
+        this.addManualAccount = this.addManualAccount.bind(this);
+
+        this.state = {
+            // One of CONST.BANK_ACCOUNT.SETUP_TYPE
+            bankAccountAddMethod: undefined,
+            hasAcceptedTerms: false,
+            routingNumber: '',
+            accountNumber: '',
+        };
+    }
+
+    toggleTerms() {
+        this.setState(prevState => ({
+            hasAcceptedTerms: !prevState.hasAcceptedTerms,
+        }));
+    }
+
+    /**
+     * @returns {Boolean}
+     */
+    canSubmitManually() {
+        return this.state.hasAcceptedTerms
+
+            // These are taken from BankCountry.js in Web-Secure
+            && CONST.BANK_ACCOUNT.REGEX.IBAN.test(this.state.accountNumber.trim())
+            && CONST.BANK_ACCOUNT.REGEX.SWIFT_BIC.test(this.state.routingNumber.trim());
+    }
+
+    addManualAccount() {
+        // @TODO call API to add account manually
+    }
+
+    render() {
+        return (
+            <View style={[styles.flex1, styles.justifyContentBetween]}>
+                <HeaderWithCloseButton
+                    title={this.props.translate('bankAccount.addBankAccount')}
+                    onCloseButtonPress={Navigation.dismissModal}
+                    onBackButtonPress={() => this.setState({bankAccountAddMethod: undefined})}
+                    shouldShowBackButton={!_.isUndefined(this.state.bankAccountAddMethod)}
+                />
+                {!this.state.bankAccountAddMethod && (
+                    <>
+                        <View style={[styles.flex1]}>
+                            <Text style={[styles.mh5, styles.mb5]}>
+                                {this.props.translate('bankAccount.toGetStarted')}
+                            </Text>
+                            <MenuItem
+                                icon={Bank}
+                                title={this.props.translate('bankAccount.logIntoYourBank')}
+                                onPress={() => {
+                                    this.setState({bankAccountAddMethod: CONST.BANK_ACCOUNT.SETUP_TYPE.PLAID});
+                                }}
+                                shouldShowRightIcon
+                            />
+                            <MenuItem
+                                icon={Paycheck}
+                                title={this.props.translate('bankAccount.connectManually')}
+                                onPress={() => {
+                                    this.setState({bankAccountAddMethod: CONST.BANK_ACCOUNT.SETUP_TYPE.MANUAL});
+                                }}
+                                shouldShowRightIcon
+                            />
+                            <View style={[styles.m5, styles.flexRow, styles.justifyContentBetween]}>
+                                <TextLink href="https://use.expensify.com/privacy">
+                                    {this.props.translate('common.privacy')}
+                                </TextLink>
+                                <View style={[styles.flexRow, styles.alignItemsCenter]}>
+                                    <TextLink
+                                        // eslint-disable-next-line max-len
+                                        href="https://community.expensify.com/discussion/5677/deep-dive-how-expensify-protects-your-information/"
+                                    >
+                                        {this.props.translate('bankAccount.yourDataIsSecure')}
+                                    </TextLink>
+                                    <View style={[styles.ml1]}>
+                                        <Icon src={Lock} fill={colors.blue} />
+                                    </View>
+                                </View>
+                            </View>
+                        </View>
+                    </>
+                )}
+                {this.state.bankAccountAddMethod === CONST.BANK_ACCOUNT.SETUP_TYPE.PLAID && (
+                    <AddPlaidBankAccount
+                        text={this.props.translate('bankAccount.plaidBodyCopy')}
+                        onSubmit={(args) => {
+                            console.debug(args);
+                        }}
+                        onExitPlaid={() => {
+                            this.setState({bankAccountAddMethod: undefined});
+                        }}
+                    />
+                )}
+                {this.state.bankAccountAddMethod === CONST.BANK_ACCOUNT.SETUP_TYPE.MANUAL && (
+                    <>
+                        <View style={[styles.m5, styles.flex1]}>
+                            <Text style={[styles.mb5]}>
+                                {this.props.translate('bankAccount.checkHelpLine')}
+                            </Text>
+                            <Image
+                                resizeMode="contain"
+                                style={[styles.exampleCheckImage, styles.mb5]}
+                                source={exampleCheckImage}
+                            />
+                            <TextInputWithLabel
+                                placeholder={this.props.translate('bankAccount.routingNumber')}
+                                keyboardType="number-pad"
+                                value={this.state.routingNumber}
+                                onChangeText={routingNumber => this.setState({routingNumber})}
+                            />
+                            <TextInputWithLabel
+                                placeholder={this.props.translate('bankAccount.accountNumber')}
+                                keyboardType="number-pad"
+                                value={this.state.accountNumber}
+                                onChangeText={accountNumber => this.setState({accountNumber})}
+                            />
+                            <CheckboxWithLabel
+                                style={[styles.mb4, styles.mt5]}
+                                isChecked={this.state.hasAcceptedTerms}
+                                onPress={this.toggleTerms}
+                                LabelComponent={() => (
+                                    <View style={[styles.flexRow, styles.alignItemsCenter]}>
+                                        <Text>
+                                            {this.props.translate('bankAccount.iAcceptThe')}
+                                        </Text>
+                                        <TextLink href="https://use.expensify.com/terms">
+                                            {`Expensify ${this.props.translate('common.termsOfService')}`}
+                                        </TextLink>
+                                    </View>
+                                )}
+                            />
+                        </View>
+                        <Button
+                            success
+                            text={this.props.translate('common.saveAndContinue')}
+                            style={[styles.m5]}
+                            isDisabled={!this.canSubmitManually()}
+                            onPress={this.addManualAccount}
+                        />
+                    </>
+                )}
+            </View>
+        );
+    }
+}
+
+BusinessBankAccountNewPage.propTypes = propTypes;
+
+export default withLocalize(BusinessBankAccountNewPage);
