+++ src/pages/ReimbursementAccount/CompanyStep.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import {View} from 'react-native';
+
+const CompanyStep = () => <View />;
+export default CompanyStep;
