+++ src/pages/ReimbursementAccount/RequestorStep.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import {View} from 'react-native';
+
+const RequestorStep = () => <View />;
+export default RequestorStep;
