+++ src/pages/ReimbursementAccount/ValidationStep.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import {View} from 'react-native';
+
+const ValidationStep = () => <View />;
+export default ValidationStep;
